Refactor: rename News-Component to NewsComponent for consistency #294
