+++ src/lib/components/NewsComponent/NewsComponent.svelte
